+++ lib/core/options/style/graph_style.dart
@@ -2,7 +2,8 @@
 //
 // This source code is licensed under Apache 2.0 License.
 
-import 'dart:ui';
+import 'dart:ui' hide TextStyle;
+import 'package:flutter/painting.dart';
 import 'dart:math' as math;
 
 import 'package:flutter_graph_view/flutter_graph_view.dart';
@@ -108,4 +109,17 @@ class GraphStyle {
   /// - `fontWeight`
   /// - `fontColor`
   VertexTextStyleGetter? vertexTextStyleGetter;
+
+  /// @en: the text style of edge.
+  /// now is only supports:
+  /// - `fontSize`
+  /// - `fontWeight`
+  /// - `fontColor`
+  ///
+  /// @zh: 顶点文字样式
+  /// 目前仅支持：
+  /// - `fontSize`
+  /// - `fontWeight`
+  /// - `fontColor`
+  EdgeTextStyleGetter? edgeTextStyleGetter;
 }