+++ lib/flutter_graph_view.dart
@@ -25,7 +25,9 @@ export 'core/options/shape/vertex_shape.dart';
 export 'core/options/shape/vertex/vertex_decorator.dart';
 export 'core/options/shape/edge_shape.dart';
 export 'core/options/shape/edge/edge_decorator.dart';
+export 'core/options/text/text_renderer.dart';
 export 'core/options/text/vertex_text_renderer.dart';
+export 'core/options/text/edge_text_renderer.dart';
 
 /// interfaces' default impl
 export 'core/algorithm/random_algorithm.dart';
@@ -56,6 +58,7 @@ export 'core/options/shape/vertex/vertex_circle_shape.dart';
 export 'core/options/shape/vertex/vertex_diamond_shape.dart';
 export 'core/options/shape/edge/edge_line_shape.dart';
 export 'core/options/text/vertex_text_renderer_impl.dart';
+export 'core/options/text/edge_text_renderer_impl.dart';
 
 /// default overlay builders
 export 'core/dashboard/k_default_overlay_builder.dart';